requirements.txt
selenium==4.18.1
google-generativeai==0.3.1
python-docx==1.1.0
requests==2.31.0

# Web Framework
Flask>=3.0.0
Flask-Login>=0.6.3
Flask-SQLAlchemy>=3.1.1
Flask-Migrate>=4.0.5
Flask-WTF>=1.2.1
WTForms>=3.1.1

# Supabase Authentication
supabase>=2.0.0
gotrue>=2.0.0

# Environment Variables
python-dotenv>=1.0.0

# Database
SQLAlchemy>=2.0.23
psycopg2-binary>=2.9.9